feat: Introduce user password update.
Introduce user password update feature. the user can update his password
by providing his email, then we check if the email is valid and send a
token code to that email address after that we check if the user provides the correct
token code that we have send via email so then if it's the correct token code the user would be able
to create a new password.